Move callbacks from Bindings into their own module.
